+++ nilearn/datasets/func.py
@@ -178,8 +178,11 @@ def fetch_nyu_rest(n_subjects=None, sessions=[1], data_dir=None, resume=True,
                    verbose=1):
     """Download and loads the NYU resting-state test-retest dataset.
 
-    For documentation see :footcite:`NYUdataset` and see :footcite:`NYUdatasetlist` for download. See :footcite:`Shehzad2009restingstate` for
-    citing the paper, and see :footcite:`ZUO20101432, ZUO20102163` for additional references.
+    For documentation see :footcite:`NYUdataset`,
+    and see :footcite:`NYUdatasetlist` for download. 
+    See :footcite:`Shehzad2009restingstate` for citing the paper, 
+    and see :footcite:`ZUO20101432`,
+    and :footcite:`ZUO20102163` for additional references.
 
     Warnings
     --------
@@ -633,7 +636,8 @@ def fetch_localizer_contrasts(contrasts, n_subjects=None, get_tmaps=False,
 
     You may cite :footcite:`PAPADOPOULOSORFANOS2017309` when using this dataset.
 
-    Scientific results obtained using this dataset are described in :footcite:`Pinel2007fast`.
+    Scientific results obtained using this dataset are described 
+    in :footcite:`Pinel2007fast`.
 
     Parameters
     ----------
@@ -1377,7 +1381,9 @@ def fetch_megatrawls_netmats(dimensionality=100, timeseries='eigen_regression',
     The network matrices are estimated from functional connectivity
     datasets of 461 subjects. Full technical details in references.
 
-    More information available in :footcite:`smithhcp2015, smith2015positive, Filippini7209, smith2014methods, reilly2009cerebellum`.
+    More information available in :footcite:`smithhcp2015`, 
+    :footcite:`smith2015positive`, :footcite:`Filippini7209`, 
+    :footcite:`smith2014methods`, and :footcite:`reilly2009cerebellum`.
 
     Parameters
     ----------