+++ doc/whats_new.rst
@@ -7,7 +7,8 @@ NEW
 Fixes
 -----
 
-- Convert references in atlas.py, func.py, neurovault.py, and struct.py to use footcite / footbibliography
+- Convert references in atlas.py, func.py, neurovault.py, and struct.py 
+  to use footcite / footbibliography.
 - Fix detrending and temporal filtering order for confounders
   in :func:`nilearn.signal.clean`, so that these operations are applied
   in the same order as for the signals, i.e., first detrending and